% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_wd.R
\name{autowd}
\alias{autowd}
\title{Set working directory function}
\usage{
autowd()
}
\description{
This function is shorthand to set the working directory to the active document's path.
It should be used as a beginning argument in a script before loading any data that's
in the same folder as the script.
}
\keyword{active}
\keyword{directory}
\keyword{document}
\keyword{path}
\keyword{set}
\keyword{working}
